Remove font-awesome link
We needed a newer version of FA, to get the Mastodon logo. The quickest and easiest way to do that was to switch to a FA "kit", hosted on their CDN. So let's remove the reference that was causing 404s in the background.

We can (and probably should) switch to a self-hosted kit in the near future, so this may switch back, but one problem at a time.