Credits fix
While I proposed the Diversity Roundtable, I really shouldn't be *on* it.  The current theory is that Jasna will be moderating (and she's welcome to add herself here as such), but for now it's mainly important to not have a cishet white guy (me) listed as being in charge of this.